requirements.txt
# ========================================
# RAG API - Production Dependencies
# ========================================
# Versões fixas para builds reproduzíveis
# Última atualização: 2025-12-26
# ========================================

# Core API Framework
fastapi==0.127.0
uvicorn[standard]==0.38.0
pydantic==2.12.4
pydantic-settings==2.12.0

# Google Cloud Platform
google-cloud-aiplatform==1.132.0
google-cloud-storage==3.6.0
google-genai==1.56.0

# Utilities
python-dotenv==1.2.1
python-multipart==0.0.20

# Authentication & Security
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4

# HTTP Client (usado por scripts de teste)
requests==2.32.5

# Testing Dependencies
pytest==8.3.4
pytest-cov==7.0.0
httpx==0.28.1

